Add AssetId property to AssetResponseDto and update Asset constructor
This commit adds a new property, AssetId, to the AssetResponseDto class. It also updates the constructor of the AssetResponseDto class to include the new AssetId parameter. Additionally, the Asset class has been updated to include the [JsonIgnore] attribute on its navigation properties. The WorkItem class has also been updated to include the [JsonIgnore] attribute on its Asset navigation property.